remove @kaze-style/react
