+++ packages/babel-plugin/src/transformPlugin.ts
@@ -17,7 +17,7 @@ type State = {
 };
 
 const options = {
-  importSource: '@kaze-style/react',
+  importSource: '@kaze-style/core',
   transforms: [
     {
       from: '__preStyle',
