+++ packages/core/README.md
@@ -14,17 +14,17 @@
   <p align="center">Kaze [風] zero-runtime CSS in JS for React<p>
   <p align="center"><b>🚧 under development 🚧</b></p>
   <div align="center">
-    <a href='https://www.npmjs.com/package/@kaze-style/react'>
-      <img src='https://img.shields.io/npm/v/@kaze-style/react?style=for-the-badge'>
+    <a href='https://www.npmjs.com/package/@kaze-style/core'>
+      <img src='https://img.shields.io/npm/v/@kaze-style/core?style=for-the-badge'>
     </a>
     <a href='https://github.com/taishinaritomi/kaze-style/blob/main/LICENSE'>
       <img src='https://img.shields.io/github/license/taishinaritomi/kaze-style?style=for-the-badge'>
     </a>
-    <a href='https://bundlephobia.com/package/@kaze-style/react'>
-      <img src='https://img.shields.io/bundlephobia/minzip/@kaze-style/react?style=for-the-badge'>
+    <a href='https://bundlephobia.com/package/@kaze-style/core'>
+      <img src='https://img.shields.io/bundlephobia/minzip/@kaze-style/core?style=for-the-badge'>
     </a>
     <a href='https://github.com/microsoft/typescript'>
-      <img src='https://img.shields.io/npm/types/@kaze-style/react?style=for-the-badge'>
+      <img src='https://img.shields.io/npm/types/@kaze-style/core?style=for-the-badge'>
     </a>
   </div>
   <br />
@@ -35,7 +35,7 @@
 - **Extract** - Can choose when to extract css is buildtime or runtime(RSC is buildtime only)
 - **Atomic** - Select atomic css with $
 - **Merge** - Style merging ignoring css specificity (atomic css only)
-- **Minimal** - [0.3kb](https://shakerphobia.netlify.app/?imports=ClassName,mergeStyle,__globalStyle,__style&pkg=@kaze-style/react) runtime by buildtime extract
+- **Minimal** - [0.3kb](https://shakerphobia.netlify.app/?imports=ClassName,mergeStyle,__globalStyle,__style&pkg=@kaze-style/core) runtime by buildtime extract
 - **TypeScript** - Type-safe styles via [csstype](https://github.com/frenic/csstype)
 - **Theme** - Consistent styling using "@kaze-style/themes"
 
@@ -43,7 +43,7 @@
 
 ```ts
 // App.style.ts
-import { createStyle, createGlobalStyle } from '@kaze-style/react';
+import { createStyle, createGlobalStyle } from '@kaze-style/core';
 
 createGlobalStyle({
   html: {
@@ -71,7 +71,7 @@ export const style = createStyle({
 
 ```ts
 // App.tsx
-import { mergeStyle } from '@kaze-style/react';
+import { mergeStyle } from '@kaze-style/core';
 import { style } from './App.style';
 
 export const App = ({ action }) => {
