+++ packages/core/src/__preStyle.ts
@@ -1,4 +1,4 @@
-import { createStyle } from './createStyle';
+import { resolveStyle } from './resolveStyle';
 import type { Classes, ForBuild } from './types/common';
 import type { KazeStyle } from './types/style';
 
@@ -8,7 +8,7 @@ export const __preStyle = <K extends string>(
   filename: string,
   index: number,
 ): Classes<K> => {
-  const [cssRules, classes, staticClasses] = createStyle(styles);
+  const [cssRules, classes, staticClasses] = resolveStyle(styles);
   if (forBuild[0] === filename) {
     forBuild[1].push(...cssRules);
     forBuild[2].push([staticClasses, index]);
