+++ packages/core/src/createGlobalStyle.ts
@@ -1,21 +1,13 @@
-import { compileNotAtomicCss } from './compileNotAtomicCss';
-import type { CssRule } from './styleOrder';
-import type { CssValue } from './types/common';
+import { resolveGlobalStyle } from './resolveGlobalStyle';
+import { setCssRules } from './setCssRules';
 import type { KazeGlobalStyle } from './types/globalStyle';
-import type { AndArray, NestObj } from './types/utils';
-import { uniqueCssRules } from './uniqueCssRules';
-
-type Result = [cssRules: CssRule[]];
 
 export const createGlobalStyle = <T extends string>(
-  _styles: KazeGlobalStyle<T>,
-): Result => {
-  const cssRules: CssRule[] = [];
-  const styles = _styles as Record<T, NestObj<AndArray<CssValue>>>;
-  for (const selector in styles) {
-    const style = styles[selector];
-    const [_cssRules] = compileNotAtomicCss(style, 'global', selector);
-    cssRules.push(..._cssRules);
+  globalStyles: KazeGlobalStyle<T>,
+): void => {
+  const [cssRules] = resolveGlobalStyle(globalStyles);
+
+  if (typeof document !== 'undefined') {
+    setCssRules(cssRules);
   }
-  return [uniqueCssRules(cssRules)];
 };
