--- packages/react/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@kaze-style/react",
-  "version": "0.15.0",
-  "license": "MIT",
-  "author": "Taishi Naritomi",
-  "description": "Kaze [風] zero-runtime CSS in JS",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/taishinaritomi/kaze-style.git",
-    "directory": "packages/react"
-  },
-  "type": "module",
-  "sideEffects": false,
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/types/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/types/index.d.ts",
-      "import": "./dist/esm/index.js",
-      "require": "./dist/cjs/index.js"
-    }
-  },
-  "files": [
-    "dist",
-    "LICENSE"
-  ],
-  "scripts": {
-    "type-check": "tsc --noEmit",
-    "exports-check": "publint",
-    "dev": "tsx ./../../scripts/build.ts --watch",
-    "build": "tsx ./../../scripts/build.ts --size"
-  },
-  "dependencies": {
-    "@kaze-style/core": "^0.15.0"
-  },
-  "peerDependencies": {
-    "react": ">=16"
-  }
-}
