--- packages/react/src/index.ts
@@ -1,17 +1,0 @@
-export * from './createStyle';
-export * from './createGlobalStyle';
-export {
-  mergeStyle,
-  __style,
-  __globalStyle,
-  __preStyle,
-  __preGlobalStyle,
-  ClassName,
-} from '@kaze-style/core';
-export type {
-  KazeStyle,
-  KazeGlobalStyle,
-  ClassNameType,
-  ClassNameRecord,
-  ClassNameOverride,
-} from '@kaze-style/core';
