--- packages/react/tsconfig.build.json
@@ -1,3 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json"
-}
