+++ playground/next-app-dir/src/app/client/page.style.ts
@@ -1,4 +1,4 @@
-import { createStyle } from '@kaze-style/react';
+import { createStyle } from '@kaze-style/core';
 
 export const style = createStyle({
   title: {
