+++ playground/next-app-dir/src/app/layout.style.ts
@@ -1,4 +1,4 @@
-import { createGlobalStyle } from '@kaze-style/react';
+import { createGlobalStyle } from '@kaze-style/core';
 import { theme } from '@kaze-style/themes';
 
 createGlobalStyle(theme.reset());
