+++ playground/next-app-dir/src/app/server/page.tsx
@@ -1,4 +1,4 @@
-import { mergeStyle } from '@kaze-style/react';
+import { mergeStyle } from '@kaze-style/core';
 import { style } from './page.style';
 import { Button } from '@/components/Button';
 import { ClientButton } from '@/components/ClientButton';
