+++ playground/next-app-dir/src/components/Button.tsx
@@ -1,5 +1,5 @@
-import type { ClassNameOverride } from '@kaze-style/react';
-import { mergeStyle } from '@kaze-style/react';
+import type { ClassNameOverride } from '@kaze-style/core';
+import { mergeStyle } from '@kaze-style/core';
 import type { ComponentProps, FC } from 'react';
 import { style } from './Button.style';
