+++ playground/next/src/components/Button.tsx
@@ -1,4 +1,4 @@
-import { mergeStyle } from '@kaze-style/react';
+import { mergeStyle } from '@kaze-style/core';
 import type { ComponentProps, FC } from 'react';
 import { style } from './Button.style';
