+++ playground/react-vite/package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@kaze-style/react": "*",
+    "@kaze-style/core": "*",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
